move acr commands to new module
