+++ aztools/luigi.py
@@ -1,4 +1,5 @@
 from shell import cmd_output
+from aztools.container_registry import acr_image_exists
 from azure.core.exceptions import ResourceNotFoundError
 from connector import Connector
 
@@ -12,16 +13,8 @@ class AzureACRImageTarget():
         self.tag = tag
         self.repository = repository
 
-    def acr_image_exists(self, repository, tag, account="fandango"):
-        """
-        checks whether a docker image exists in an azure ACR repository
-        """
-        cmd = f"az acr repository show-tags  --name {account} --resource-group fandango --repository {repository}"
-        output = cmd_output(cmd)
-        return f'"{tag}"' in output
-
     def exists(self):
-        return self.acr_image_exists(self.repository, self.tag)
+        return acr_image_exists(self.repository, self.tag)
 
 
 class AzureBlobTarget():
